fixup! GApplication: Add RestartData property, and setter
This significantly reworks the previous commit. I’ve kept it as a
separate commit for now to ease review. Before merging, it should be
squashed into the original commit though.

Significant changes in this commit:
 - Change from using a property to using a build/consume interface. This
   means we’re not forced to keep the restart data in memory all the
   time. That’s helpful, because it might be quite large. It also means
   that the data is immutable, which should avoid a whole class of bugs
   related to mutating it badly.
 - Add more documentation.

Work still to do:
 - Unit tests (once people are happy with the API).

See the dependent work in GTK and gnome-session-manager:
 - https://gitlab.gnome.org/pwithnall/gtk/-/tree/save-state
 - https://gitlab.gnome.org/pwithnall/gnome-session/-/tree/session-restore

At first glance, it might look like exposing the `RestartData` on the
`org.gtk.Application` D-Bus interface means that all the secret data
inside an app is now world-readable. However, this is fine within the
desktop’s security model: user processes are not isolated from each
other. They could just as easily open the saved session data from the
user’s home directory. Flatpak applications are isolated from each other
on D-Bus by default, so this is not a problem for them.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>